web/requirements.txt
asgiref==3.3.1
coverage==5.5
Django==3.1.7
django-cors-headers==3.7.0
django-markdownx==3.0.1
djangorestframework==3.12.2
Markdown==3.3.4
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1
